README.md
this is a website resembling a blog.
there is an admin which can add posts, edit them, and delete them. admin also has access to all the comments and only accepted 
comments by the admin can be shown in the website.
every user can sign up by entering a username, an email, and a password.
and then they can sign in to their account if they log out.
every user can add a post, but can only edit & delete their own posts.
any user can search for a phrase in the search bar, and can see the posts by the order of date(sign up/in not required).